Fixed the "Who initiated the connection?" bug
There's no way to know which device initiates the Bluetooth connection.
So we need to keep our data ready to send. For that, BluetoothManager
needs to know which data to access, i.e. it needs to know the distance
range. So we need to send that itself in the JSONObject. Now the
MainActivity keeps track of all the data to be sent along with the help
of the InfoPool.