kinda sorta hacky way to get Twitter data
!TESTED
Launching the browser instance in SettingsActivity seems to destroy and
recreate the MainActivity and thus everything that's in it, including
the TwitterListFragment. So I've had to remove clearing the
SharedPreferences in the Twitter tab's onDestroy(), which is also called
just before recreating rather than on launching the browser activity as
expected (by me anyway).